Move components to be top level
